Move genericService to its own file
